Extract optional steps into a component
