+++ src/components/styles/OptionalStepComponent.module.css
@@ -1,0 +1,37 @@
+.container {
+  margin-bottom: 1.5rem;
+}
+
+.enableSection {
+  margin-bottom: 1.5rem;
+  padding: 1rem;
+  background-color: var(--ifm-color-emphasis-100);
+  border-radius: 4px;
+}
+
+.enableCheckbox {
+  font-weight: bold;
+  display: flex;
+  align-items: center;
+  cursor: pointer;
+}
+
+.infoText {
+  margin-top: 0.5rem;
+  margin-bottom: 0;
+  color: var(--ifm-color-emphasis-700);
+}
+
+.sectionTitle {
+  font-size: 1.2rem;
+  font-weight: 600;
+  margin: 1rem 0 1rem;
+  padding-bottom: 0.5rem;
+  border-bottom: 1px solid var(--ifm-color-emphasis-200);
+}
+
+.instanceCount {
+  display: block;
+  margin-bottom: 1rem;
+  font-weight: 500;
+}
